[REFACTOR] check-value.js
